Slightly better Allwinner CedarX support (override IVE, IOS, IFBM, IVBV)
Introducing the missing libve_set_ive(), libve_set_ios(),
libve_set_ifbm(), libve_set_ivbv() functions, we can now
override IVE, IOS, IFBM, IVBV interfaces.

The CedarX VLC (with the added use of the above mentioned
functions) at least does not segfault now, but still fails
and produces the following error messages:

Warning: There are not more bitstream!
Warning: The fifo is full!
Warning: There are not more bitstream!
Warning: The fifo is full!
Warning: There are not more bitstream!
Warning: The fifo is full!
Warning: There are not more bitstream!
Warning: The fifo is full!
Warning: There are not more bitstream!
Warning: The fifo is full!
Warning: There are not more bitstream!
Warning: The fifo is full!
Warning: There are not more bitstream!
Warning: The fifo is full!
Warning: There are not more bitstream!
Warning: The fifo is full!

Now we need to add detailed tracing of all the calls in order to
debug this issue and verify/tweak ABI compatibility.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>